# kanal/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kanal API",
    "version": "0.0.0"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "kn_live_<base62> | kn_test_<base62>"
      }
    },
...
  },
  "paths": {
...
    "/v1/secrets": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "secrets": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "purpose": {
                            "type": "string",
                            "enum": [
                              "byok_anthropic",
                              "byok_google",
                              "byok_openai",
                              "wa_token",
                              "notion_token",
                              "slack_token",
                              "generic_oauth",
                              "signing_secret"
                            ]
                          },
                          "metadata": {
                            "type": "object",
                            "additionalProperties": {},
                            "nullable": true
                          },
                          "kmsKeyId": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "rotatedAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          }
                        },
                        "required": [
                          "id",
                          "purpose",
                          "metadata",
                          "kmsKeyId",
                          "createdAt",
                          "rotatedAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "secrets"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "purpose": {
                    "type": "string",
                    "enum": [
                      "byok_anthropic",
                      "byok_google",
                      "byok_openai",
                      "wa_token",
                      "notion_token",
                      "slack_token",
                      "generic_oauth",
                      "signing_secret"
                    ]
                  },
                  "value": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 10000
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                },
                "required": [
                  "purpose",
                  "value"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "purpose": {
                      "type": "string",
                      "enum": [
                        "byok_anthropic",
                        "byok_google",
                        "byok_openai",
                        "wa_token",
                        "notion_token",
                        "slack_token",
                        "generic_oauth",
                        "signing_secret"
                      ]
                    },
                    "metadata": {
                      "type": "object",
                      "additionalProperties": {},
                      "nullable": true
                    },
                    "kmsKeyId": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "rotatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "purpose",
                    "metadata",
                    "kmsKeyId",
                    "createdAt",
                    "rotatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/secrets/{id}/rotate": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "value": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 10000
                  }
                },
                "required": [
                  "value"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "purpose": {
                      "type": "string",
                      "enum": [
                        "byok_anthropic",
                        "byok_google",
                        "byok_openai",
                        "wa_token",
                        "notion_token",
                        "slack_token",
                        "generic_oauth",
                        "signing_secret"
                      ]
                    },
                    "metadata": {
                      "type": "object",
                      "additionalProperties": {},
                      "nullable": true
                    },
                    "kmsKeyId": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "rotatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "purpose",
                    "metadata",
                    "kmsKeyId",
                    "createdAt",
                    "rotatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/secrets/{id}": {
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "enum": [
                    "null"
                  ],
                  "nullable": true
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.kanal.app"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ]
}